Adding toleration field to the cluster registration mutation
